Implementation configuration

+++ src/main/java/ch/kalunight/zoe/command/delete/DeletePlayerCommand.java
@@ -6,9 +6,10 @@ import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import ch.kalunight.zoe.ServerData;
 import ch.kalunight.zoe.command.CommandUtil;
-import ch.kalunight.zoe.model.Player;
 import ch.kalunight.zoe.model.Server;
-import ch.kalunight.zoe.model.SpellingLangage;
+import ch.kalunight.zoe.model.config.ServerConfiguration;
+import ch.kalunight.zoe.model.player_data.Player;
+import ch.kalunight.zoe.model.static_data.SpellingLangage;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.Member;
@@ -20,30 +21,46 @@ public class DeletePlayerCommand extends Command {
 
   public DeletePlayerCommand() {
     this.name = USAGE_NAME;
-    this.help = "Delete the given player. Manage Channel permission needed.";
+    this.help = "Delete the given player.";
     this.arguments = "@DiscordMentionOfPlayer";
-    Permission[] permissionRequired = {Permission.MANAGE_CHANNEL};
-    this.userPermissions = permissionRequired;
     this.helpBiConsumer = getHelpMethod();
   }
 
   @Override
   protected void execute(CommandEvent event) {
     event.getTextChannel().sendTyping().complete();
-    Server server = checkServer(event.getGuild());
-
+    
+    Server server = ServerData.getServers().get(event.getGuild().getId());
+    
+    if(!server.getConfig().getUserSelfAdding().isOptionActivated() && !event.getMember().getPermissions().contains(Permission.MANAGE_CHANNEL)) {
+        event.reply("You need the permission \"" + Permission.MANAGE_CHANNEL.getName() + "\" to do that.");
+        return;
+    }
+    
     List<Member> members = event.getMessage().getMentionedMembers();
 
     if(members.size() != 1) {
       event.reply("You need to mention 1 people !");
     } else {
       User user = members.get(0).getUser();
+      
+      if(!user.equals(event.getAuthor()) && !event.getMember().getPermissions().contains(Permission.MANAGE_CHANNEL)) {
+        event.reply("You cannot delete another player than you if you don't have the *" + Permission.MANAGE_CHANNEL.getName() + "* permission.");
+        return;
+      }
+      
       Player player = server.getPlayerByDiscordId(user.getId());
-
+      
       if(player == null) {
         event.reply(user.getName() + " is not registered !");
       } else {
         server.deletePlayer(player);
+        if(server.getConfig().getZoeRoleOption().getRole() != null) {
+          Member member = server.getGuild().getMember(user);
+          if(member != null) {
+            server.getGuild().getController().removeRolesFromMember(member, server.getConfig().getZoeRoleOption().getRole()).queue();
+          }
+        }
         event.reply(player.getDiscordUser().getName() + " has been deleted !");
       }
     }
@@ -53,7 +70,7 @@ public class DeletePlayerCommand extends Command {
     Server server = ServerData.getServers().get(guild.getId());
 
     if(server == null) {
-      server = new Server(guild, SpellingLangage.EN);
+      server = new Server(guild, SpellingLangage.EN, new ServerConfiguration());
     }
     return server;
   }
